+++ CHANGELOG.md
@@ -6,5 +6,36 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Megaphone [1.0.0] 
+- Added missing network asset
+
+## Megaphone [0.1.8] 
+- Updated readme
+
+## Megaphone [0.1.6]~[0.1.7] 
+
+- Added death SFX when the player dies while using it
+- Added modes : low pitch, high pitch, siren
+- Added more configs
+
+## Megaphone [0.1.5] 
+
+- Mode now synced between clients
+- Mode now saved per item, allowing multiple items to be used
+
+## Megaphone [0.1.4] 
+
+- Added more configurations
+- Updated readme
+
+## Megaphone [0.1.3] 
+
+- Added 'Loud voice' filter mode
+- Added 3D base model
+
+## Megaphone [0.1.1] 
+
+- Audio effects components now added when item equipped or used
+
+## Megaphone [0.1.0] 
 
 - Initial commit